Minor fixes (#2480)
* Update common_gui.py

Little fix to the `validate_model_path` function to properly handle a folder path. Currently, it errors out when a folder path is selected for a training with diffusers. 😊

* Update lora_gui.py

Switching the validation type for the resume training state path from 'file' to 'folder'.
😊

* Update common_gui.py

format

* Update common_gui.py

My last fix was wrong and was returning errors when using a default model or when the VAE was not defined. This new fix works with all possibilities.
However, when a diffusers folder is used, `validate_file_path` will show a failed check, but the operation will still succeed because `validate_folder_path` will return successfully.