docs: doc uv install method is fine for Docker use
- Update the uv installation instructions removing Docker from the list of exceptions and affected notes.

We have now switched to using uv for building docker image.
Both uv and pip are effective options for the docker environment, and I believe there is no need to highlight either method when making a selection.
I have left the Runpod as is because I do not have experience with Runpod.

Signed-off-by: CHEN, CHUN <jim60105@gmail.com>